Added new instructions

+++ aoc2019rs/src/solutions/day2.rs
@@ -4,7 +4,7 @@ use crate::utils;
 use crate::intcode::IntcodeMachine;
 
 pub fn run() -> Result<String> {
-    let program = utils::read_input_list_as::<usize>(2, b',')?;
+    let program = utils::read_input_list_as::<i64>(2, b',')?;
     
     // let part_1_result = run_test(&program, 12, 2)?;
 
@@ -21,7 +21,7 @@ pub fn run() -> Result<String> {
     bail!("Couldn't find inputs with output 19690720")
 }
 
-fn run_test(program: &[usize], noun: usize, verb: usize) -> Result<usize> {
+fn run_test(program: &[i64], noun: i64, verb: i64) -> Result<i64> {
     let mut machine = IntcodeMachine::new(&program);
     machine.write_memory(1, noun);
     machine.write_memory(2, verb);
